feature: `with` individual type remapping support for Rust
Add individual type remapping in the with parameter of the wit_bindgen::generate macro

The type remapping happens in the `InterfaceGenerator::type_path` only for the types provided via with macros parameter. Effectively, the remapped type is used in place of the original type everywhere in the generated bindings for any imports and exports, including lifting and lowering functions. The original type definition is not generated.
This approach puts a requirement on the remapped type to have the same internal structure and identical to what would wit-bindgen generate (including alignment, etc.), since lifting/lowering uses its fields directly.
See the `codegen::with_type` test for the example of the remapping in action for various types.