rust: Store raw intrinsics in vtables
Instead of generating a wrapper-per-intrinsic try to instead store raw
functions inside of vtables to push as much code as possible into `mod
async_support`, as it's generally easier to grok code in one location
rather than generated code.

Note that this does not update the `read` or `write` vtable functions as
they're going to be trickier, if even possible, to expose the more raw
version.

This shouldn't have any functional change, it's just a reorganizing.
The `cancel_*` intrinsics are all marked with a `TODO`, however, to come
back and reconsider the return value which is otherwise discarded right
now.