Add a `wit-bindgen test` subcommand
As I've read more and more of the Rust async runtime support and other
various bits and pieces I've wanted more and more the ability to easily
write tests for guest interactions with the host. While I don't think
it's feasible to generate arbitrary hosts I do think it's possible to do
this much more easily than is done today with the testing support in
this repository. In essence this commit is an implementation of #1161.

The goal of this commit is to add a `wit-bindgen test` test suite
runner. This test suite will be used to migrate all existing tests in
this repository to this new framework. In the limit this is expected to
make it easier to write tests (no Rust knowledge necessary), make it
more flexible to write tests (now you can use raw `*.wat`), and
additionally improve the quality of the test suite by making it more
reusable. The reusability isn't the highest priority at this time as
it's not clear what else would want to reuse this, but my hope is that
this refactoring is at least a large-ish leap forward towards having a
component model test suite of some kind eventually.