Update support for p3 to match the latest wasm-tools (#1198)
* Update wasm-tools dependencies

* Fixes from wasm-tools update to p3 support

Fix the name and signature of various intrinsics in the Rust bindings
generator for p3 support. This is detected by updating wasm-tools in the
wasip3-prototyping repository which requires changes here.

* minimal stub of waitable-set support

This is temporary (and relies on the host to ignore the waitable-set parameter)
until we have proper waitable-set support, which I'll work on soon.

Signed-off-by: Joel Dice <joel.dice@fermyon.com>

---------

Signed-off-by: Joel Dice <joel.dice@fermyon.com>
Co-authored-by: Joel Dice <joel.dice@fermyon.com>