Add a `wit-bindgen test` subcommand (#1192)
* Add a `wit-bindgen test` subcommand

As I've read more and more of the Rust async runtime support and other
various bits and pieces I've wanted more and more the ability to easily
write tests for guest interactions with the host. While I don't think
it's feasible to generate arbitrary hosts I do think it's possible to do
this much more easily than is done today with the testing support in
this repository. In essence this commit is an implementation of #1161.

The goal of this commit is to add a `wit-bindgen test` test suite
runner. This test suite will be used to migrate all existing tests in
this repository to this new framework. In the limit this is expected to
make it easier to write tests (no Rust knowledge necessary), make it
more flexible to write tests (now you can use raw `*.wat`), and
additionally improve the quality of the test suite by making it more
reusable. The reusability isn't the highest priority at this time as
it's not clear what else would want to reuse this, but my hope is that
this refactoring is at least a large-ish leap forward towards having a
component model test suite of some kind eventually.

* Add the p2 target

* Add new crate to publish list

* Update crate metadata

* Add back in crate test

* Don't prepare ignored languages

* Actually fail the process if tests fail

* Install wasmtime on CI too

* Always install wasi-sdk for all deps (which is now just a bunch of deps
too)

* Migrate some more Rust codegen tests to the new framework

* Improve rendering of errors in test subcommand

Use one deduplicated method for showing a list of errors.

* Remove warnings in generated Rust programs

* More migrated Rust codegen tests

Add in configuration to drop default bindings args, aka `--generate-all`
in Rust.

* Move Rust/C-specific tests to their own folder

* Migrate another rust codegen test

* Migrate Rust numbers test to new framework

* Migrate numbers/wasm.c to new framework

* Migrate the `wasm/lists.rs` test to the new framework

* Migrate `lists/wasm.c` test to new framework

* Update test docs

* Beef up Rust support to pull in `futures` and support WASIp1

* The `futures` crate will be necessary for upcoming async tests.
* Support for WASIp1 is required as `wasm32-wasip2` can't produce an
  async component natively as it's not new enough.
* Test runner support is extended slightly to support passing runtime
  arguments, such as `-Wcomponent-model-async`.

* Document test's `config.rs`

* Enable debuginfo in Rust tests

Helps symbolicate backtraces better.

* Fixup some merge conflicts