rust: Flag futures/streams as owned handles (#1210)
This commit updates how bindings work for futures/streams in Rust by
ensuring that the type information for them encodes that they behave
like `own<T>` handles in that passing the value to an import for
examples passes ownership of the value in bindings. This ensures that
generates APIs appropriately model relinquishing ownership of the
`FutureReader<T>` end, for example.