Migrate the `rust_xcrate` test to `wit-bindgen test` (#1236)
* Migrate the `rust_xcrate` test to `wit-bindgen test`

This was a tricky test to migrate. Doing this required implementing a
number of new features:

* Rust now has a `[lang]`-specific configuration `externs = [...]` which
  is a list of files to compile as external crates. This tests the
  cross-crate behavior of the test.
* Tests now support more than just runner/test composed components, now
  they also support multiple components being composed. This is required
  because the test was specifically testing behavior where a component
  simultaneously imports and exports.

This last point was particularly tricky. Namely this required
abstracting the definition of a test to a runner plus a list of tests,
not just one test. These are still modeled as a list of `world` items in
the original WIT document. Additionally I was having trouble getting
`wasm-compose` working so I ended up switching to `wac`. This means that
tests can now support custom `wac` composition scripts to configure how
exactly the composition is created, and that's used in this new test to
insert a component as a sandwich between two others.

* Review comments