Refactor how imported tasks and bindings work  (#1260)
* Abstract `Waitable` and its code

* Refactor how imported tasks and bindings work

This commit refactors imported tasks to work the same way as
futures/streams using a `WaitableOperation<S>` for managing all state
associated with the task. This involved refactoring generating bindings
and redoing how imported tasks are represented. This notably creates a
"hole" for cancellation to slot into in the near future and additionally
addresses a number of issues around leaking memory in async imports upon
cancellation.

Additionally to get tests passing in wasip3-prototyping it was necessary
to fix an issue where a stack-local pointer was deallocated with a
heap-deallocation function. Fixing this required refactoring how
exported tasks worked which involved siloing abstractions from each
other and restructuring the flow of information throughout
`FutureState`. In a sense these are refactorings I wanted to get around
to anyway, but they were required to be here as well to pass upstream tests.

* Update crates/core/src/abi.rs

Co-authored-by: Joel Dice <joel.dice@fermyon.com>

---------

Co-authored-by: Joel Dice <joel.dice@fermyon.com>