Migrate `resources` test to `wit-bindgen test`
I've started mimicking the pattern of some of the more complicated
resource tests where I copy the exact test files as before and configure
the test to not need any modifications there. Not exactly the easiest
way to write the test and probably testing too much but it least makes
these diffs smaller and/or easier to make.