Start running some async tests in CI (#1286)
This commit starts to execute some async tests in CI via artifacts
downloaded from the wasip3-prototyping repository. While the status is
reported on PRs this won't actually result in failing to merge something
if the job fails at this time since it's known that async is a bit of a
shifting landscape. In the meantime it's hoped that at least having this
status on PRs will be sufficient for seeing failures for now.

The actual set of initial tests here are relatively edge-case-y where I
added them some time ago when testing Rust bindings. They're not
comprehensive by any means but are hopefully at least a start. I hope to
next start adding more tests in follow-up PRs.