Add some more async tests (#1287)
* Add some more async tests

I went through various parts of the guest runtime and sprinkled
`if true { panic!() }` throughout. Anything that wasn't hit got a test.
This led to a few issues in wasip3-prototyping:

* https://github.com/bytecodealliance/wasip3-prototyping/issues/137
* https://github.com/bytecodealliance/wasip3-prototyping/issues/138
* https://github.com/bytecodealliance/wasip3-prototyping/issues/139
* https://github.com/bytecodealliance/wasip3-prototyping/issues/140

so not all tests are fully enabled yet but they should get enabled as
bugs get fixed.

Additionally this starts to use `CALLBACK_CODE_YIELD`. That's disabled
for callbacks due to bytecodealliance/wasip3-prototyping#140 but it's
enabled for `block_on` where it translates to a `waitable-set.poll`
instead of a `waitable-set.wait`.

* Fix fallback build of `poll`

* Adjust poll signature

* Fix native signature