Support "linear writes" in Rust `future` bindings (#1306)
To account for WebAssembly/wasip3-prototyping#172 and
WebAssembly/component-model#521 the Rust bindings need to take into
account that a future writer cannot be dropped without actually writing
something. To model this futures now must be created with a thunk that
creates the default value to be sent. This value is only used if another
value was not otherwise written.

Closes #1304