update `future.{write,read}` ABIs
This updates the Rust generator to use the new ABI defined in
https://github.com/WebAssembly/component-model/pull/524.  The C generator will
need a similar update as well.

- `future.write` now accepts its payload value as up to 4 flat parameters before spilling to linear memory.

- `future.read` takes no payload pointer when it has no payload type

- `{stream,future}.close-{readable,writable}` have been renamed to `{stream,future}.drop-{readable,writable}`

I've tested this end-to-end with `wasip3-prototyping` and will soon post a
corresponding PR there.

Finally, `rustfmt` was complaining about using `gen` as a field name, so I took
this opportunity to address that.

Signed-off-by: Joel Dice <joel.dice@fermyon.com>