update `future`/`stream` ABIs and names
This updates `wit-bindgen-rt` etc. to use (some of) the new ABI defined in
https://github.com/WebAssembly/component-model/pull/524.  It covers everything
in that PR _except_ the lifting and lowering changes to `future.{read,write}`,
which we decided are more trouble than they're worth.

I've tested this end-to-end with `wasip3-prototyping` and will soon post a
corresponding PR there.

Finally, `rustfmt` was complaining about using `gen` as a field name, so I took
this opportunity to address that.

Signed-off-by: Joel Dice <joel.dice@fermyon.com>

fix rust codegen build errors

Signed-off-by: Joel Dice <joel.dice@fermyon.com>

snapshot

Signed-off-by: Joel Dice <joel.dice@fermyon.com>