fix(core): result check for async function
This commit fixes a check that was incorrectly looking at the CM level
function in the case of an async guest import when determining whether
it had the right results.

The problem with the existing logic is that it was trying to check for
a constraint on the wrong function -- the guest import being lowered
rather than the core function underneath that.

The case that makes this obvious was something like an `run: async
func()` where the CM level function clearly has no result, but the
underlying core function being async lifted must of course have a i32
status code as a result.