Parse `wit` folder in `generate!` with `inline`
This commit updates handling of the `inline` attribute in the
`generate!` macro of the Rust bindings generator. Previously if `inline`
was specified it would by default lookup nothing on the filesystem
unless `path` was specified. This contradicts the documentation though
that `./wit` is the default lookup path. To resolve this this commit
adds a check to see if `./wit` exists, and if so parses it. This keeps
the default search path behavior without requiring its existence when
using `inline`.