Prioritize `inline` worlds, not external ones (#1358)
When multiple `path`s are used the `generate!` macro will fail if they
all contain a `world` and no world is otherwise specified. This then
additionally affected when using `inline` plus loading auxiliary WIT
from a `paths` directory (either explicitly or the default `wit`). This
commit switches the logic to prioritizing worlds in the `inline` block
because if that's being specified it's pretty likely the exact one that
wants to be used.