Redefine `with` in the Rust generator to only apply to imports
This commit avoids looking at `--with` for exports, for example, to fix
situations where an interface is imported and exported and only the
import types should be reused from elsewhere. I'm not aware of any cases
where exports want to be reused, so there's effectively no more support
for that.