Support lists of WIT directories on the CLI. (#1368)
Accept multiple WIT paths in the wit-bindgen CLI, using the same logic
as the multiple WIT paths accepted by the `path` parameter in the
Rust bindgen macro.

Also, document that list-of-WIT interfaces, in the Rust macro and now
the wit-bindgen CLI, have a topological ordering requirement which may
be lifted in the future.