Move `wit_bindgen::spawn` behind a feature (#1392)
* Move `wit_bindgen::spawn` behind a feature

I do not know of a great solution for #1305 so instead of actually
solving it at this time the next-best thing I can think of is to
guide users towards not using it of sorts. This commit adds an
`async-spawn` feature to the `wit-bindgen` crate which gates this
function and the feature is disabled by default. This notably also
enables putting the `futures` crate behind this feature which should
greatly reduce the compile-time-impact of taking on this dependency
(`futures` is a bit large). Refactoring was required internally to fully
sever the `futures` dependency except for spawning.

* Fix doc check by using nightly

* Fix a feature name