Generalize bindings for futures  (#1409)
* Add more mutability to internal async traits

As usage expands a bit beyond just this crate it can be useful to have
mutability in the guts of the implementation. It's also easy enough to
provide this everywhere so it's not much of a burden to plumb it around.

* Generalize bindings for futures

This commit is similar to #1390 in that the goal is to enable external
systems to use the Rust bindings here without requiring the exact shape
of the vtable structure and such. This is, for example, intended to be
useful when working on `wit-dylib`.