Rely less on `wasm-component-ld` for tests (#1434)
* Rely less on `wasm-component-ld` for tests

This repository is often used to test experimental/upcoming features for
components such as async/etc, but this creates a chicken-and-egg problem
where it becomes difficult to compile tests. Many tests use upstream
toolchains which use `wasm-component-ld` (e.g. Rust/C tests) but it can
take time to get this updated. There's already logic to handle having
tests target wasip1 and running the wit-component step manually, so this
PR makes that step unconditional. This means that test binaries are
always assembled with a crate-dependency of `wit-component` which should
enable easier updates since it's "just" a Cargo dependency vs being part
of a toolchain in a separate project.

* Conditionally embed type information