go: restore handles on incomplete writes
We now keep track of any resource/stream/future handles we've lowered while
writing to a stream or future and restore any that were unwritten so they can be
used (e.g. possibly written) again.

While I was working on this, clippy pointed out that `std::io::pipe` (which I
started using in a previous commit) was added in Rust 1.87, so I've bumped the
MSRV to match.

Also, the test case I added to cover this revealed another bug, which I've fixed
here: we weren't generating valid code for async functions which return tuples.

Fixes #1458

Signed-off-by: Joel Dice <joel.dice@fermyon.com>