remove `CallbackCode::Poll` (#1471)
* remove `CallbackCode::Poll`

As of https://github.com/WebAssembly/component-model/pull/578, that code no
longer exists.  Instead, we can emulate it via a combination of
`CallbackCode::Yield` and `waitable-set.poll`.

I'm working on a PR to make the corresponding changes to Wasmtime, and needed
this to make the tests pass.

* update C and Go generators to eliminate CALLBACK_CODE_POLL

See my earlier commit updating the Rust generator for details.

* fix test comment typos